+++ src/configUI.ts
@@ -18,6 +18,7 @@ export const enum EDiagMasterSwitch {
 }
 
 type TempConfigs = {
+    autoSwitchAhk2: boolean;
     statusBarDisplayColor: string;
     formatTextReplace: boolean;
     baseScan: {
@@ -58,6 +59,7 @@ function getConfigs<T>(section: string): T {
 
 function getConfig(): TConfigs {
     const ed: TConfigs = {
+        autoSwitchAhk2: getConfigs<boolean>('autoSwitchAhk2'),
         statusBarDisplayColor: getConfigs<string>('statusBar.displayColor'),
         formatTextReplace: getConfigs<boolean>('format.textReplace'),
         baseScan: {
@@ -177,6 +179,10 @@ export function needDiag(): boolean {
         return false;
     }
     return true;
+}
+
+export function isAutoSwitchAhk2(): boolean {
+    return config.autoSwitchAhk2;
 }
 // vscode.window.setStatusBarMessage(timeSpend);
 // vscode.window.showErrorMessage()